[dev-tools] support for ESM repos
