--- .prettierrc.js
@@ -1,11 +1,0 @@
-const {getPrettierConfig, deepMerge} = require('ocular-dev-tools');
-
-const defaultConfig = getPrettierConfig({react: '16.8.2'});
-
-// Make any changes to default config here
-const config = deepMerge(defaultConfig, {});
-
-// Uncomment to log the eslint config
-// console.debug(config);
-
-module.exports = config;
