--- .yarnrc
@@ -1,6 +1,0 @@
-# THIS IS AN AUTOGENERATED FILE. DO NOT EDIT THIS FILE DIRECTLY.
-# yarn lockfile v1
-
-
-lastUpdateCheck 1624384382671
-yarn-path ".yarn/releases/yarn-1.22.10.js"
