--- modules/dev-tools/node/get-puppeteer-executable-path.js
@@ -1,23 +1,0 @@
-const fs = require('fs');
-const altPath = process.argv[2];
-
-let executablePath = getExecutablePath();
-
-if (!executablePath && altPath) {
-  executablePath = getExecutablePath(altPath);
-}
-
-console.log(executablePath);
-
-function getExecutablePath(dir) {
-  try {
-    const puppeteer = require(dir ? `${dir}/node_modules/puppeteer` : 'puppeteer');
-    const path = puppeteer.executablePath();
-    if (fs.existsSync(path)) {
-      return path;
-    }
-  } catch (err) {
-    // ignore
-  }
-  return null;
-}
