--- modules/dev-tools/node/module-dir.js
@@ -1,6 +1,0 @@
-const {resolve} = require('path');
-
-// Prints the installed location of ocular-dev-tools
-module.exports = function () {
-  console.log(resolve(__dirname, '..'));
-};
