+++ modules/dev-tools/package.json
@@ -3,6 +3,7 @@
   "description": "Dev tools for our Javascript frameworks",
   "license": "MIT",
   "version": "2.0.0-alpha.2",
+  "type": "module",
   "keywords": [
     "javascript"
   ],
@@ -13,14 +14,16 @@
   "files": [
     "src",
     "config",
-    "node",
     "scripts",
     "templates",
-    "test",
     "CHANGELOG.md"
   ],
+  "exports": {
+    ".": "./index.js",
+    "./configuration": "./src/configuration/index.cjs"
+  },
   "types": "./src/index.d.ts",
-  "main": "./src/index.js",
+  "main": "./src/index.cjs",
   "bin": {
     "ocular-bootstrap": "./scripts/bootstrap.js",
     "ocular-build": "./scripts/build.js",
@@ -97,6 +100,7 @@
     "typescript": "~4.6.0",
     "ts-node": "~10.9.0",
     "tsconfig-paths": "^4.1.1",
+    "url": "^0.11.0",
     "vite": "^4.0.1",
     "vite-plugin-html": "^3.2.0"
   },
@@ -115,5 +119,8 @@
   },
   "peerDependencies": {
     "math.gl": "^3.0.0"
+  },
+  "engines": {
+    "node": ">= 14.8"
   }
 }
