+++ modules/dev-tools/scripts/lint.sh
@@ -7,14 +7,14 @@ set -e
 # lint.sh DIR1,DIR2
 MODE=$1
 
-DEV_TOOLS_DIR=`node -e "require('ocular-dev-tools/node/module-dir')()"`
+DEV_TOOLS_DIR=$(dirname $0)/..
 
-DIRECTORIES=`node $DEV_TOOLS_DIR/node/get-config.js ".lint.paths"`
+DIRECTORIES=`node $DEV_TOOLS_DIR/src/helpers/get-config.js ".lint.paths"`
 if [[ $DIRECTORIES == *","* ]]; then
   DIRECTORIES={$DIRECTORIES}
 fi
 
-EXTENSIONS=`node $DEV_TOOLS_DIR/node/get-config.js ".lint.extensions"`
+EXTENSIONS=`node $DEV_TOOLS_DIR/src/helpers/get-config.js ".lint.extensions"`
 if [[ $EXTENSIONS == *","* ]]; then
   EXTENSIONS={$EXTENSIONS}
 fi
