+++ modules/dev-tools/scripts/test.js
@@ -1,9 +1,8 @@
 #!/usr/bin/env node
 
-const shell = require('shelljs');
-const path = require('path');
+import shell from 'shelljs';
+import {join} from 'path';
 
+const scriptDir = new URL(import.meta.url).pathname;
 // Runs the bash script and forward the arguments, exiting with the same code
-shell.exit(
-  shell.exec(`${path.resolve(__dirname, './test.sh')} ${process.argv.slice(2).join(' ')}`).code
-);
+shell.exit(shell.exec(`${join(scriptDir, '../test.sh')} ${process.argv.slice(2).join(' ')}`).code);
