+++ modules/dev-tools/src/index.js
@@ -1,22 +1,1 @@
-const {getOcularConfig} = require('./helpers/get-ocular-config');
-
-// JS Tool Configurations
-const {getBabelConfig} = require('./configuration/get-babel-config');
-const {getESLintConfig} = require('./configuration/get-eslint-config');
-const {getPrettierConfig} = require('./configuration/get-prettier-config');
-
-// Utilities
-const deepMerge = require('deepmerge');
-
-module.exports = {
-  // Ocular Configuration
-  getOcularConfig,
-
-  // JS Tool Configurations
-  getBabelConfig,
-  getESLintConfig,
-  getPrettierConfig,
-
-  // Utilities
-  deepMerge
-};
+export {getOcularConfig} from './helpers/get-ocular-config';
