+++ modules/dev-tools/test/lib/utils.spec.js
@@ -1,5 +1,5 @@
 import test from 'tape-promise/tape';
-import {shallowMerge} from '../../src/utils/utils';
+import {shallowMerge} from '../../src/utils/utils.js';
 
 test('dev-tools#utils', (t) => {
   t.equals(typeof shallowMerge, 'function');
