+++ tsconfig.json
@@ -1,6 +1,7 @@
 {
   "compilerOptions": {
     "target": "es2020",
+    "module": "esnext",
     "jsx": "react",
     "strict": true,
     "noImplicitAny": false,
